refactor: move CSS analytics contracts to output crate
